--- src/components/forms/index.ts
@@ -1,10 +1,0 @@
-// Form components
-export * from "./fields";
-export * from "./sections";
-
-// Re-export common types from transformers to avoid conflicts
-export type {
-  SelectOption,
-  StringSelectOption,
-  CheckboxOption,
-} from "@/utils/formDataTransformers";
