+++ src/components/forms/register/ProgressBar.tsx
