+++ src/components/ui/form.tsx
@@ -8,7 +8,7 @@ import {
   useFormContext,
 } from "react-hook-form";
 import { Label } from "@/components/ui/label";
-import { cn } from "@/components/lib/utils";
+import { cn } from "@/lib/utils";
 
 const Form = FormProvider;
