+++ src/components/ui/label.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import * as React from "react";
-import { cn } from "@/components/lib/utils";
+import { cn } from "@/lib/utils";
 
 // Updated interface with optional children
 interface LabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {
