+++ src/components/ui/sidebar.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import { cn } from "@/components/lib/utils";
+import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
 import {
   User,
