+++ src/components/ui/tabs.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as TabsPrimitive from "@radix-ui/react-tabs";
 
-import { cn } from "@components/lib/utils";
+import { cn } from "@/lib/utils";
 
 const Tabs = TabsPrimitive.Root;
