+++ src/components/ui/toast.tsx
@@ -5,7 +5,7 @@ import * as ToastPrimitives from "@radix-ui/react-toast";
 import { cva, type VariantProps } from "class-variance-authority";
 import { X } from "lucide-react";
 
-import { cn } from "@components/lib/utils";
+import { cn } from "@/lib/utils";
 
 const ToastProvider = ToastPrimitives.Provider;
